Now getStopCfg.sh downloads config files for 2016,2017,2018. Only make supplementary file softlinks if the files exist.

+++ Tools/scripts/getStopCfg.sh
@@ -24,12 +24,16 @@ OPTIND=1    # Reset in case getopts has been used previously in the shell.
 SUPP_CFG=supplementaryFiles.cfg
 SETS_CFG_2016=sampleSets_PostProcessed_2016.cfg
 SETS_CFG_2017=sampleSets_PostProcessed_2017.cfg
+SETS_CFG_2018=sampleSets_PostProcessed_2018.cfg
 COLL_CFG_2016=sampleCollections_2016.cfg
 COLL_CFG_2017=sampleCollections_2017.cfg
+COLL_CFG_2018=sampleCollections_2018.cfg
 SETS_LINK_NAME_2016=$SETS_CFG_2016
 SETS_LINK_NAME_2017=$SETS_CFG_2017
+SETS_LINK_NAME_2018=$SETS_CFG_2018
 COLL_LINK_NAME_2016=$COLL_CFG_2016
 COLL_LINK_NAME_2017=$COLL_CFG_2017
+COLL_LINK_NAME_2018=$COLL_CFG_2018
 SUPP_FILE_DIR=supplementaryFiles
 TARBALL=SUPP_FILE_TARBALL.tar.gz
 
@@ -43,8 +47,10 @@ function print_help {
     echo "    -d checkout_directory :  This is the directory where the configuration files will be downloaded to (default: .)"
     echo "    -a SETS_LINK_NAME_2016 : Specify this option to name the softlink to \"$SETS_CFG_2016\" something other than \"$SETS_LINK_NAME_2016\""
     echo "    -b SETS_LINK_NAME_2017 : Specify this option to name the softlink to \"$SETS_CFG_2017\" something other than \"$SETS_LINK_NAME_2017\""
+    echo "    -c SETS_LINK_NAME_2018 : Specify this option to name the softlink to \"$SETS_CFG_2018\" something other than \"$SETS_LINK_NAME_2018\""
     echo "    -x COLL_LINK_NAME_2016 : Specify this option to name the softlink to \"$COLL_CFG_2016\" something other than \"$COLL_LINK_NAME_2016\""
     echo "    -y COLL_LINK_NAME_2017 : Specify this option to name the softlink to \"$COLL_CFG_2017\" something other than \"$COLL_LINK_NAME_2017\""
+    echo "    -z COLL_LINK_NAME_2018 : Specify this option to name the softlink to \"$COLL_CFG_2018\" something other than \"$COLL_LINK_NAME_2018\""
     echo "    -o :                     Overwrite the softlinks if they already exist"
     echo "    -n :                     Download files without producing softlinks"
     echo "    -v :                     increase verbosity: print more stuff... for those who like stuff"
@@ -74,9 +80,13 @@ while getopts "h?t:d:a:b:x:y:onv" opt; do
         ;;
     b)  SETS_LINK_NAME_2017=$OPTARG
         ;;
+    c)  SETS_LINK_NAME_2018=$OPTARG
+        ;;
     x)  COLL_LINK_NAME_2016=$OPTARG
         ;;
     y)  COLL_LINK_NAME_2017=$OPTARG
+        ;;
+    z)  COLL_LINK_NAME_2018=$OPTARG
         ;;
     o) OVERWRITE="-f"
         ;;
@@ -236,8 +246,10 @@ then
     echo " - Creating softlinks to $REPO_NAME config files"
     ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2016 $SETS_LINK_NAME_2016 > /dev/null 2>&1 
     ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2017 $SETS_LINK_NAME_2017 > /dev/null 2>&1 
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2018 $SETS_LINK_NAME_2018 > /dev/null 2>&1 
     ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2016 $COLL_LINK_NAME_2016 > /dev/null 2>&1
     ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2017 $COLL_LINK_NAME_2017 > /dev/null 2>&1
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2018 $COLL_LINK_NAME_2018 > /dev/null 2>&1
     # make list of files to tar
     if [[ -f $DOWNLOAD_DIR/$SUPP_CFG ]]
     then
@@ -245,7 +257,10 @@ then
         while read -r line || [[ -n "$line" ]]
         do
             SUPP_FILE="$line"
-            ln $OVERWRITE -s $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE $SUPP_FILE > /dev/null 2>&1
+            if [ -f $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE ]
+            then
+                ln $OVERWRITE -s $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE $SUPP_FILE > /dev/null 2>&1
+            fi
         done < $DOWNLOAD_DIR/$SUPP_CFG
     else
         echo "ERROR: Config file $DOWNLOAD_DIR/$SUPP_CFG not found"
